add adjusted triaxial test to the test_element_lab
